- dev와 병합 및 이미지 업로드 로직 수정.

+++ src/main/kotlin/com/challengeteamkotlin/campdaddy/presentation/product/dto/request/CreateProductRequest.kt
@@ -8,7 +8,7 @@ import org.springframework.web.multipart.MultipartFile
 data class CreateProductRequest(
     val title:String,
     val content:String,
-    val pricePerDay : Int,
+    val pricePerDay : Long,
     val images:List<String>,
     val category: String,
